Consolidate small clusters tominimize number of test batches
This commit adds consolidation of test batches, subject to
maxClusterSize, using the "first fit descending" algorithm. I.e.,
clusters are sorted by descending size, and are added to the first
bin into which they fit. If no bin fits, a new one is created. This
is proven to be within 22% of the minimum possible number of bins.